poa_agents/tier1_validation_agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "poa-tier1-validation-agent"
version = "0.1.0"
description = "Tier 1 deterministic validation agent for POA applications"
requires-python = ">=3.12"
dependencies = [
    "agentex-sdk",
    "scale-gp",
    "openai>=1.0.0",
    "azure-identity",
    "supabase>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["project"]

